backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
python-dotenv
requests
pydantic-settings
PyJWT
passlib[bcrypt]
email-validator
